LayoutTests/imported/w3c/web-platform-tests/html/semantics/popovers/popover-focus-2-expected.txt

Button1  Button2  Invoker1  Button3  Button4
Invoker  after
Show popover Hide popover

FAIL Popover focus navigation assert_equals: Hidden popover should be skipped expected Element node <button id="button2">Button2</button> but got Element node <body><div id="fixup">
  <button id="button1">Button1</bu...
FAIL Circular reference tab navigation assert_equals: Step 2 expected Element node <button id="circular1" autofocus="" popovertarget="popove... but got Element node <body><div id="fixup">
  <button id="button1">Button1</bu...
FAIL Popover focus returns when popover is hidden by invoker assert_equals: Hide button should be focused due to autofocus attribute expected Element node <button popovertarget="deleted1" popovertargetaction="hid... but got Element node <button popovertarget="deleted1" popovertargetaction="sho...
